Add Key and Door for Childs room
TODO: Fix script not deleting key from inventory menu when it should.
See if I can optimize the door so that I don't have to change the box collider base depending on if it's a automatic door or not. (If not and collider is large, player can't select door when too close.)